plugins/ppr-beacon/pyproject.toml
[tool.poetry]
name = "ppr-beacon"
version = "0.1.0"
description = "SEO-optimized static mini-sites for verified food pantry locations"
authors = ["For The Greater Good"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "^3.1"
psycopg2-binary = "^2.9"
python-slugify = "^8.0"
structlog = "^24.0"
boto3 = "^1.34"
pydantic = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"
ruff = "^0.4"
mypy = "^1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "S", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
